v2

+++ src/main/java/br/com/infnet/minddesk/controllers/SolicitacaoController.java
@@ -1,6 +1,7 @@
 package br.com.infnet.minddesk.controllers;
 
 import br.com.infnet.minddesk.model.*;
+import br.com.infnet.minddesk.services.DTOConverterService;
 import br.com.infnet.minddesk.services.impl.SolicitacaoServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -18,41 +19,15 @@ public class SolicitacaoController {
     @Autowired
     private SolicitacaoServiceImpl solicitacaoService;
 
+    @Autowired
+    private DTOConverterService dtoConverterService;
+
     @PostMapping
-    public ResponseEntity<Solicitacao> criarSolicitacao(@RequestBody Solicitacao solicitacao) {
-        // Criar um novo Agente com base nos dados fornecidos no JSON
-        Agente agente = new Agente();
-        agente.setId(solicitacao.getAgente().getId());
-        agente.setNome(solicitacao.getAgente().getNome());
-        agente.setEmail(solicitacao.getAgente().getEmail());
-        agente.setCargo(solicitacao.getAgente().getCargo());
-
-        // Criar uma nova Categoria com base nos dados fornecidos no JSON
-        Categoria categoria = new Categoria();
-        categoria.setNome(solicitacao.getCategoria().getNome());
-        categoria.setId(solicitacao.getCategoria().getId());
-        categoria.setDescricao(solicitacao.getCategoria().getDescricao());
-
-        // Criar um novo Cliente com base nos dados fornecidos no JSON
-        Cliente cliente = new Cliente();
-        cliente.setNome(solicitacao.getCliente().getNome());
-        cliente.setEmail(solicitacao.getCliente().getEmail());
-        cliente.setTelefone(solicitacao.getCliente().getTelefone());
-        cliente.setId(solicitacao.getCliente().getId());
-        cliente.setSolicitacoes(solicitacao.getCliente().getSolicitacoes());
-
-        // Definir os agentes, artigos, categoria e cliente na solicitação
-        solicitacao.setAgente(agente);
-        solicitacao.setCategoria(categoria);
-        solicitacao.setCliente(cliente);
-
-        // Salvar a solicitação com os dados atualizados
+    public ResponseEntity<Solicitacao> criarSolicitacao(@RequestBody Solicitacao json) {
+        Solicitacao solicitacao = dtoConverterService.converterParaSolicitacaoDTO(json);
         solicitacaoService.save(solicitacao);
-
-        // Retornar a solicitação criada com o status 201 (Created)
         return ResponseEntity.status(HttpStatus.CREATED).body(solicitacao);
     }
-
 
     @GetMapping
     public ResponseEntity<List<Solicitacao>> listarSolicitacoes() {
@@ -61,14 +36,14 @@ public class SolicitacaoController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Solicitacao> buscarSolicitacaoPorId(@PathVariable UUID id) {
+    public ResponseEntity<Solicitacao> buscarSolicitacaoPorId(@PathVariable Long id) {
         Optional<Solicitacao> solicitacaoOptional = solicitacaoService.findById(id);
         return solicitacaoOptional.map(ResponseEntity::ok)
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Solicitacao> atualizarSolicitacao(@PathVariable UUID id, @RequestBody Solicitacao solicitacaoAtualizada) {
+    public ResponseEntity<Solicitacao> atualizarSolicitacao(@PathVariable Long id, @RequestBody Solicitacao solicitacaoAtualizada) {
         Optional<Solicitacao> solicitacaoOptional = solicitacaoService.findById(id);
         if (solicitacaoOptional.isEmpty()) {
             return ResponseEntity.notFound().build();
@@ -79,8 +54,13 @@ public class SolicitacaoController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletarSolicitacao(@PathVariable UUID id) {
-        solicitacaoService.deleteById(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<String> deletarSolicitacao(@PathVariable Long id) {
+        Optional<Solicitacao> solicitacaoOptional = solicitacaoService.findById(id);
+        if (solicitacaoOptional.isPresent()) {
+            solicitacaoService.deleteById(id);
+            return ResponseEntity.ok("Solicitação deletada com sucesso");
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
